Moved fragments and activity to view folder
