+++ app/src/main/res/layout-land/activity_main.xml
@@ -27,7 +27,7 @@
 
     <fragment
         android:id="@+id/fragment_navigation_drawer"
-        android:name="com.example.gm_challenge.ElementFragment"
+        android:name="com.example.gm_challenge.view.fragment.ElementFragment"
         android:layout_width="@dimen/nav_drawer_width"
         android:layout_height="0dp"
         android:layout_gravity="start"
