+++ app/src/main/res/layout-sw900dp/activity_main_tablet.xml
@@ -29,7 +29,7 @@
 
     <fragment
         android:id="@+id/fragment_navigation_drawer"
-        android:name="com.example.gm_challenge.ElementFragment"
+        android:name="com.example.gm_challenge.view.fragment.ElementFragment"
         android:layout_width="@dimen/nav_drawer_width"
         android:layout_height="0dp"
         android:layout_gravity="start"
